requirementts.txt
git+https://github.com/huggingface/transformers.git
git+https://github.com/huggingface/accelerate.git
git+https://github.com/huggingface/diffusers.git
git+https://github.com/huggingface/peft.git
transformers-stream-generator
huggingface_hub
optimum-quanto
sentencepiece
opencv-python
torchvision
safetensors
xformers
requests
hf_xet
spaces
pillow
gradio
einops
torch
timm